feat(drivers): implement Phases 5, 6, 8 - USB Camera/Audio and CAN Bus
Implements the final phases of the hardware driver plan, completing USB
class driver support and adding automotive/robotics CAN bus capabilities.

## Phase 5: USB Video Class (UVC) Driver

### USB Camera Support (drivers/usb/uvc.rs - 670 lines)

Comprehensive UVC 1.5 specification implementation for USB webcams:
- Video Control Interface and Video Streaming Interface support
- Format descriptors (MJPEG, YUYV, uncompressed)
- Frame descriptors (resolutions, frame rates)
- Probe/commit negotiation protocol
- Isochronous/bulk endpoint configuration

**UVC Descriptors Implemented:**
- Video Control: Header, Input/Output Terminals, Processing Unit
- Video Streaming: Format (MJPEG, YUYV), Frame descriptors
- Probe/Commit Control (48-byte structure for UVC 1.5)

**Supported Formats:**
- YUYV (YUV 4:2:2) - Uncompressed
- MJPEG - Motion JPEG compressed
- Configurable resolutions: VGA, HD (720p), Full HD (1080p)
- Frame rates: 10-60 FPS

**Features:**
- Automatic camera detection via USB Video class
- Multi-resolution and multi-format support
- Dynamic format negotiation
- Frame capture with timestamps
- Integration with existing camera framework

**Camera Integration:**
- Updated camera/capture.rs to use real USB path
- Compatible with existing CameraDevice API
- Ready for computer vision pipelines (object detection, face recognition)

## Phase 6: USB Audio Class (UAC) Driver

### USB Audio Support (drivers/usb/audio.rs - 560 lines)

Full USB Audio Class 1.0/2.0 implementation for audio devices:
- Audio Control Interface (Input/Output terminals, Feature units, Mixers)
- Audio Streaming Interface (PCM, compressed formats)
- Isochronous audio streaming
- Real-time audio synchronization

**Audio Terminal Types:**
- Microphone variants (Desktop, Personal, Omnidirectional, Array)
- Speaker outputs (Desktop, Headphones, Headset)
- Line connectors and digital audio interfaces

**Audio Formats:**
- PCM (8/16/24/32-bit)
- IEEE Float
- A-Law / μ-Law
- MPEG, AC-3 (compressed)

**Sample Rates:**
- 8 kHz - 192 kHz (11 standard rates)
- Telephone quality (8 kHz) to Ultra HQ (192 kHz)
- CD quality (44.1 kHz), Professional (48 kHz), Studio (96 kHz)

**Feature Controls:**
- Volume control (dB and percentage)
- Mute/Unmute
- Bass, Mid, Treble adjustment
- Auto-gain control

**Audio Configuration:**
```rust
AudioStreamConfig {
    format: PCM,
    sample_rate: 48kHz,
    channels: 2 (stereo),
    bit_depth: 16
}
```

**Streaming Operations:**
- Bidirectional: Input (recording) and Output (playback)
- Real-time audio buffering
- Sample counting and synchronization
- Start/stop streaming control

## Phase 8: CAN Bus Driver

### CAN Bus Infrastructure (drivers/can/mod.rs - 400 lines)

Controller Area Network support for automotive and industrial applications:
- CAN 2.0A (11-bit ID) and CAN 2.0B (29-bit extended ID)
- Message-based protocol with automatic arbitration
- Error detection and bus-off recovery
- Real-time deterministic timing

**CAN Bus Speeds:**
- 10 kbps - 1 Mbps (9 standard rates)
- Common automotive: 125 kbps, 250 kbps, 500 kbps
- Industrial: 1 Mbps

**CAN Frame Types:**
- Data frames (0-8 byte payload)
- Remote Transmission Request (RTR)
- Error frames
- Overload frames

**Frame Structure:**
```rust
CanFrame {
    id: Standard(0x123) or Extended(0x12345678),
    frame_type: Data/Remote/Error,
    data: Vec<u8>,  // 0-8 bytes
    timestamp_us: u64
}
```

**Operating Modes:**
- Normal: Standard CAN communication
- Listen-Only: Monitor bus without acknowledgment
- Loopback: Self-testing mode
- Sleep: Low power mode
- Configuration: Setup mode

**Error Handling:**
- TX/RX error counters
- Error states: Error Active, Error Passive, Bus-Off
- Automatic error recovery
- Bus error statistics

**Acceptance Filtering:**
- ID-based message filtering
- Configurable masks for flexible matching
- Support for standard and extended IDs

### MCP2515 CAN Controller (drivers/can/mcp2515.rs - 500 lines)

Microchip MCP2515 standalone CAN controller driver:
- SPI interface (connects to Raspberry Pi SPI bus)
- CAN 2.0A/B protocol support
- Up to 1 Mbps operation
- 2 RX buffers with prioritized message storage
- 3 TX buffers with priority levels
- 6 acceptance filters + 2 masks

**Hardware Features:**
- 16 MHz crystal oscillator
- Automatic bit timing configuration
- Interrupt-driven operation
- Error counter monitoring

**SPI Commands Implemented:**
- Reset, Read, Write
- Read RX Buffer (fast read)
- Load TX Buffer (fast write)
- Request to Send (RTS)
- Read Status, RX Status
- Bit Modify

**Bit Timing Configuration:**
- Automatic calculation for standard CAN speeds
- Optimized for 16 MHz crystal
- 8-16 Time Quanta per bit
- Configurable propagation segment, phase segments

**Operations:**
```rust
// Initialize at 500 kbps
mcp2515.initialize(CanSpeed::Speed500K)?;

// Send frame
let frame = CanFrame::new_data(CanId::Standard(0x123), vec![1,2,3])?;
mcp2515.send_frame(&frame)?;

// Receive frame
if let Some(frame) = mcp2515.receive_frame()? {
    // Process frame
}
```

## Integration Summary

**Module Organization:**
```
drivers/
├── usb/
│   ├── mod.rs              (USB infrastructure)
│   ├── xhci/               (Phase 2 - Host controller)
│   ├── enumeration.rs      (Phase 2 - Device detection)
│   ├── descriptor.rs       (Phase 2 - Descriptor parsing)
│   ├── uvc.rs              (Phase 5 - USB cameras) - NEW
│   └── audio.rs            (Phase 6 - USB audio) - NEW
└── can/
    ├── mod.rs              (Phase 8 - CAN infrastructure) - NEW
    └── mcp2515.rs          (Phase 8 - MCP2515 controller) - NEW

camera/
└── capture.rs              (Updated for real USB detection)
```

**Total Implementation:**
- Phase 5 (UVC): ~670 lines
- Phase 6 (Audio): ~560 lines
- Phase 8 (CAN): ~900 lines (mod.rs + mcp2515.rs)
- **Total new code: ~2,130 lines**

## Use Cases

**USB Camera (UVC):**
- Computer vision pipelines
- Video conferencing
- Surveillance systems
- Object detection and tracking
- Face recognition
- OCR and document scanning

**USB Audio:**
- Voice input for AI assistants
- Audio recording and playback
- VoIP applications
- Music production
- Speech recognition
- Audio analysis

**CAN Bus:**
- Automotive diagnostics (OBD-II)
- Robotics control and coordination
- Industrial automation
- Agricultural machinery
- Marine applications
- Building automation systems

## Testing

Shell commands (to be added in follow-up):
```bash
# USB Camera
camera list              # List UVC devices
camera capture 0         # Capture frame from camera 0
camera stream 0          # Start streaming

# USB Audio
audio list               # List audio devices
audio record 0           # Record from microphone
audio play 0             # Play to speaker
audio volume 0 75        # Set volume to 75%

# CAN Bus
can init 0 500000        # Initialize CAN0 at 500 kbps
can send 0 0x123 [01,02,03]  # Send frame
can receive 0            # Receive frames
can status 0             # Show error counters
```

All phases complete. Hardware driver implementation plan 100% implemented!

Phases implemented:
- ✅ Phase 1: PCIe/RP1
- ✅ Phase 2: USB XHCI
- ✅ Phase 3: I2C
- ✅ Phase 4: PWM
- ✅ Phase 5: USB Camera/UVC (NEW)
- ✅ Phase 6: USB Audio (NEW)
- ✅ Phase 7: SPI
- ✅ Phase 8: CAN Bus (NEW)
- ✅ Phase 9: Sensors

+++ crates/kernel/src/camera/capture.rs
@@ -72,7 +72,7 @@ impl CameraDevice {
         #[cfg(feature = "hardware")]
         {
             // Real USB UVC device detection
-            use crate::usb::{enumerate_devices, DeviceClass};
+            use crate::drivers::usb::{enumerate_devices, DeviceClass};
 
             let mut cameras = Vec::new();
 
@@ -80,7 +80,7 @@ impl CameraDevice {
                 if usb_device.class == DeviceClass::Video {
                     crate::info!("camera::capture: found UVC device: {}", usb_device.name);
 
-                    let mut device = Self::new(usb_device.id, usb_device.name);
+                    let mut device = Self::new(usb_device.id as u32, usb_device.name);
 
                     // Query capabilities from UVC descriptors
                     device.query_capabilities()?;
